Swith from roboto to ubuntu font
We don't want to use or download robot but use the ubuntu font on our page.
For this, we need to replace paper-styles-classes.html which downloads the
roboto font and put them in non overridable classes. This is what the new
paper-styles-without-roboto-classes.html does.

Commenting as well for now paper-toast which erronously import
paper-styles/typography.html which then download roboto font. Will open an
upstream PR for this.