Add a loading animation for home page
The initial home page loading is animation all categories one after another
from bottom to final position.
We trigger that animation manually on first page loading (and thus, we have to
set it invisible first, and then trigger visibility + animation on when dom is
ready). Then, as we use the "entry" keyword for this animation, next page
routing will trigger it (through neon-animated-pages).