Proper animation between page and a background.
As neon-animation-pages set its children to position absolute, we need to
inflate somehow the page children's size to have background element (colored)
fitting content without having fixed size. For that, use a wrapper div and
adapt rules (like no padding) for the wrapper div.

Also handle that heros have a very big index contrary to the rest of the page.
Here, we have a background hero (until we implement ripple effects), so change
the z-index.