Set padding on non full screen page to internal wrapper
Move and then import that padding from shared-styles.
We should be able at some point to put that on the host, to avoid a div
wrapper.

We can't set it on the page itself for all pages as before as we have some
pages that are fullscreen with a background color.