Minimize the number of grid computation
What happened is that many computed properties are parameters for other
properties. The issue is that for each affected computed property by the same
value, this trigger multiple times recomputed the leaf properties even if the
transaction is not over.
For now, we only depend on fewer properties and call functions directly as much
as possible.